requirements.txt
google-api-python-client==1.9.3
google-cloud-storage==1.33.0
